fix: cross type copying [bugfix] (CT-1496) (#6614)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CT-1496**

### Brief description. What is this change?
After the hack-week refactor speak and text steps, (as well as no match no reply sections), regardless if on a voice/chat project, both use the same contents for the prompt:
https://github.com/voiceflow/creator-app/blob/7d10d169bc3d168fa8c8f008d6eab5b373616c90/packages/creator-app/src/components/Prompt/index.tsx#L11-L22

Which is the active project type. So even if it is a speak step, if you are on a chat project, it will try to render the contents of the speak as a text, and this causes a crash.

For now this just prevents people from copying across steps from a chat type project to a voice type project. This is unfortunate if you are copying common steps like logic stuff.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>