fix: remove import from file (CT-000) (#6596)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CT-000**

### Brief description. What is this change?
To fix the importing large projects issue, we are now using an HTTP endpoint with creator-api to import it, and then just broadcasting the event on realtime.

But we still had to keep the old implementation because many users may have out-of-date clients. 
It's been some time since the fix is released, hopefully we can remove it soon. Let's wait a little longer before merging this PR.

https://github.com/voiceflow/creator-app/pull/6573

Co-authored-by: Tyler Han <tylerhan97@gmail.com>