feat: transcript timestamp based on update at rather than created at (PL-363) (#6531)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-363**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

Changes the Transcript UI to use the `updatedAt` time instead of `createdAt` to timestamp the transcripts. 

This is necessary because in the MS Teams use-case, because the chat is a continuous history of all messages between the user and the bot ever made. Thus, the transcripts will get updated asynchronously with recent user interactions, rather than being an immutable record that is created once (as it is with transcript's original use-case). 

The Teams transcript could have been created months or years in the past, and so, the `createdAt` timestamp does not reflect the fact that the transcript has been recently updated, becoming more useless over time.