feat: load workspace in the workspace channel [bugfix] (BUG-349) (#6884)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements BUG-349**

### Brief description. What is this change?

since we have a DB sync issue (the user can be a workspace member according to new tables and not according to old ones), so there's an edge case. When we load workspaces from the identity and creator-api and mapping creator-api workspace (set plan/seats/etc) onto identity one, sometimes we can't fully map workspaces, cause according to old tables user is not part of the workspace. So seats, plans, and other creator-api info are undefined in the creator-app. This fix ensures that when we connect to the workspace channel we refetch the entire workspace, the single workspace get method uses auth for permissions so the user is able to fetch the workspace from the old table.

- load workspace in the workspace channel
- remove the board list from the workspace since it never been used
- other cleanups 

### Checklist

- [ ] changes have been validated in an ephemeral environment
- [ ] this is a breaking change and should publish a new major version
- [ ] appropriate tests have been written
- [ ] any new env vars have been added to the [notion doc](https://www.notion.so/voiceflow/Add-Environment-Variables-be1b0136479f45f1adece7995a7adbfb) and infra has been notified