fix: lines not being rendered when going to a specific node (CT-1572) (#6898)
**Fixes or implements BUG-347**
**Fixes or implements CT-1572**

### Brief description. What is this change?
When moving to a component or subtopic lines were not rendered. That's because some race condition where the diagram is not fully rendered and we try to focus the node. Added a small timeout of `100ms` and it works. Not sure if thats the best solution tho, but it fixes the issue for now and i don't see any side effect adding that.