fix: overlapping/duplicate topics (BUG-524) (#7113)
I think this is some kind of crazy race condition.
The only place where this can happen is on reorder. I've simulated some randomness by adding random `await`s:
<img width="817" alt="Screenshot 2023-08-04 at 11 32 48 AM" src="https://github.com/voiceflow/creator-app/assets/5643574/fc2fa437-746e-4513-9c44-372c86f34daf">

In theory it is possible if you are reordering the same element, to remove it twice, then add it twice. (The second reorder happens before the first one is complete)

The only way to prevent this is to have a stricter filter check on addTopics, it is more atomic when it is on the mongo side. Only add the element if it DNE already. `$addToSet` doesn't work and we have to use `$push` because we need `$position`

Co-authored-by: Tyler Han <tylerhan97@gmail.com>