fix: remove prototype webhook (CT-000) (#7140)
What is the prototype webhook? This was originally a feature built for BMW, but not known or used by anyone else.

How it works is when you call an API endpoint, it will then send a request to your prototype and display a message. I've commented where this core logic was (and it is janky).

![Screenshot 2023-08-25 at 7 38 29 AM](https://github.com/voiceflow/creator-app/assets/5643574/8984d6a9-9a6f-4407-9d71-7c8a910855f7)



Co-authored-by: Tyler Han <tylerhan97@gmail.com>