fix: new markdown library (BUG-555) (#7146)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements BUG-555**

### Brief description. What is this change?
switching over to `react-markdown`

`markdown-to-jsx` has a ton of problems that freezes can crashes people's browsers:
https://github.com/probablyup/markdown-to-jsx/issues/496
https://github.com/probablyup/markdown-to-jsx/issues/473
https://github.com/probablyup/markdown-to-jsx/issues/438

The source code on that library is just too much to go through. 

The alternative is `react-markdown` with some plugins. I've had to update the serializer to work with it:
https://github.com/voiceflow/slate-serializer/commit/bac5460fda388e37d942a15a6ddff94370cc381a
https://github.com/voiceflow/react-chat/commit/5f59262bc55687e63460f1402868c001164e0d7f


I've gone through and tried my best to make most if not all the styling consistent. Did a lot of testing there.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>