fix: remove google sheets and zapier (CT-000) (#7169)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CT-000**

### Brief description. What is this change?
Removing all frontend support of `zapier` and `google sheets`. The integrations step === the API step now
This has been communicated to the growth team. The steps will continue to function on `general-service` and `general-runtime` (if you publish a new version, it will still work, you just can't edit or access the steps anymore).

The most important thing is to ensure that the API step still works as intended. I've done testing for this.

Reference comments in the code.

BEFORE:
![Screenshot 2023-09-06 at 4 20 59 PM](https://github.com/voiceflow/creator-app/assets/5643574/532c550c-9760-47b1-8df6-10ac2b2d751e)

AFTER:
![Screenshot 2023-09-06 at 4 35 44 PM](https://github.com/voiceflow/creator-app/assets/5643574/0896190c-66c6-4db7-bc9f-c065ad195026)

The Fail port and link will be missing, but I promise it's still working in the backend


Co-authored-by: Tyler Han <tylerhan97@gmail.com>