fix: optional sources for faq (VF-000) (#7226)
We are adding a beta feature for FAQs in the KB which are may not be associated with documents, meaning the sources are empty. 

This change prevents the frontend from completely breaking if a KB response does not have a source file name.

This is also a temporary fix for the corrupted data bug, where deleting a file while processing would cause the frontend to crash.

As a side effect, there may be views where the source is empty. The only times this will happen is if the alternative was a frontend crash. As we move from a beta to a full on FAQ feature, we will be able to add sources and names to all KB objects including FAQs. For now however, we will use this approach.

Side effect (empty second source):
<img width="256" alt="Screenshot 2023-10-04 at 10 00 24 AM" src="https://github.com/voiceflow/creator-app/assets/23105545/1c1b2cb2-1149-479e-891a-b2d4af27ff04">

Side effect (no name for KB preview chunk):
<img width="511" alt="Screenshot 2023-10-04 at 10 05 23 AM" src="https://github.com/voiceflow/creator-app/assets/23105545/baa57ca9-baac-4483-a766-6c57b5313245">


Co-authored-by: Erik Vandenberg <erikvandenberg81@gmail.com>