fix: fake project list upsert (CT-000) (#7258)
`upsert` is busted with soft deletes. There's no proper way to do it without writing raw SQL, I looked EXTENSIVELY through mikro-orm and knex. 

The table `app_cxd.workspace_project_lists` is not 1 to 1 with workspaces - I think in the future we will have other tables like this. So some form of an upsert operation is needed.

I had to write `'run'` because it was clearly not well documented/typed by Mikro:
https://github.com/mikro-orm/mikro-orm/blob/54f4d026939eccfcfb9c9ed23617659c3d513073/packages/postgresql/src/PostgreSqlConnection.ts#L30-L45

You only get `affectedRows` IF you put in `'run'` (default is `all`). But the types make it look like you always get that as a response.

Everything else is self explanatory. I've tested this thorougly and have spent an embarrassing number of hours trying to make it work. There is no better way to do this besides raw SQL. I know it is possible for a race condition. 

Co-authored-by: Tyler Han <tylerhan97@gmail.com>