fix: byo llm (CV3-515) (#7273)
a bit of a revision on https://github.com/voiceflow/creator-app/pull/7268 after discussing with @Diophontine 
reused as much logic as I could.

We'll use an environment variable to dictate the available models rather than a feature flag.
This will be more scalable in case Cisco wants both GPT 3.5 and GPT 4 in the future. Also a core ask is that the model is very obviously denoted as a custom model. So by using `CLOUD_ENV` it will say "Cisco GPT 3.5".

Co-authored-by: Tyler Han <tylerhan97@gmail.com>