fix: lower cache interval (CT-000) (#7288)
So it turns out if you leave a tab on the side, chrome will throttle all timers and intervals to 1 minute:
https://stackoverflow.com/questions/64836377/what-is-chromiums-throttle-javascript-timers-in-background-feature-and-how-to

Our DIAGRAM HEARTBEAT is hard coded on the FE to 7 seconds, but this will turn into 1 minute on an inactive tab. Through a lot of testing, I've found that because the heartbeat starts a little bit (let's say ~2 seconds) AFTER the permissions/cache is set (cache is set on initial page load, heartbeat starts when diagram is fully loaded).

This means under usual circumstances, the heartbeat will HIT the cache, and then ~2 seconds later the cache is expired. This means that there is roughly a minute interval now where the cache is not set.
The next heartbeat will create a new cache, but it will be this on/off cycle. So at any given point you can assume around half our users who are tabbed out don't have a cached auth.

When we restart realtime, it will broadcast an event and force everyone without a cache to call auth **at the same time**.
This is my best hypothesis as to how to reduce a little bit of load for the auth service.

![image](https://github.com/voiceflow/creator-app/assets/5643574/d87f5bc2-d0da-4c61-8a38-d910d83b4f14)
![image](https://github.com/voiceflow/creator-app/assets/5643574/836c3deb-a054-4dc9-87e7-58b9bf69ba16)


Co-authored-by: Tyler Han <tylerhan97@gmail.com>