fix: prototype check format with dtos (CT-000) (#7320)
add `@voiceflow/dtos` to `creator-app`
I had to change the format of the `TraceDTOFactory` to give better typing. Now every time before we show something on the prototype, we have to validate that the JSON is actually correct, not just the type.

This is so that we can allow users to add their own "custom" version of things.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>