fix: bump all logux instances to latest version (CT-000) (#7329)
we have a lot of different random versions of logux around this monorepo.

Just consolidating everything to the latest versions. I tried my best to check the log and ensure no breaking changes.
This isn't a bugfix or anything. It's just for consolidating and easier debugging in the future.

`ml-gateway`, `socket-utils`, `creator-app` all have different versions than `realtime` which is on the latest.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>