fix: project list permissions [bugfix] (CT-000) (#7337)
editors don't have access to `WORKSPACE_UPDATE`.
It is only reserved for admins and owners. 

But editors should be able to shuffle projects around on our legacy projectLists screen. The permission they have is to `create a project` so we're going to use that.

I've also updated `removeProjectFromList` to use `Permission.PROJECT_DELETE`

Co-authored-by: Tyler Han <tylerhan97@gmail.com>