fix: base concurrency on actual machine (CT-000) (#7393)
default concurrency on turborepo is 10:
![image](https://github.com/voiceflow/creator-app/assets/5643574/1e529f01-416c-4a09-b18b-4208210540c0)

Vercel docs:
https://turbo.build/repo/docs/reference/command-line-reference/run#--concurrency

We have to explicitly state 100% for it to run the correct amount. 
- Medium Docker CircleCIs only have 3 cores
- Large Docker CircleCIs only have 4 cores.

there's no way for turbo the detect the number of "vCPUs" allocated to it, so we will just hard code things for now.
The actual machine is always much bigger. i.e.:
<img width="850" alt="Screenshot 2023-11-22 at 3 07 40 PM" src="https://github.com/voiceflow/creator-app/assets/5643574/fda4c1e0-399f-4b92-9d65-069a2a9bed08">

The machine says it has 36CPUs and 18 per socket, but in reality we only get 4. So yeah.

Turborepo uses this rust crate to detect CPUs: https://docs.rs/num_cpus/latest/num_cpus/ I ran it inside an SSH container job. 
<img width="450" alt="Screenshot 2023-11-22 at 3 27 06 PM" src="https://github.com/voiceflow/creator-app/assets/5643574/1bc9b377-503b-4cca-96ce-1dbbbddc8725">



Co-authored-by: Tyler Han <tylerhan97@gmail.com>