feat: backups frontend (CV3-673) (#7411)
**Fixes or implements CV3-673**

### Brief description. What is this change?
<img width="1391" alt="Screenshot 2023-11-27 at 5 34 10 PM" src="https://github.com/voiceflow/creator-app/assets/5643574/f456cdd4-29d2-45b0-827e-56737e065b9e">

Added a new endpoint to realtime to fetch environments. I've tested it just in case `liveVersion` or `devVersion` points to nothing. Today it is hard coded and there is only `production` and `development`

I've made the Environments page a full branch from the Versions page.

I've also update all the styling for the tables so they all reference the old version, that way they are not duplicated 3 times. Hopefully we can just remove everything soon.

Remaining work:
* convert version to backup
* add a confirm delete modal for deleting versions
* have a write up and docs, right now it's just lorem ipsum
* preview backup

Co-authored-by: Tyler Han <tylerhan97@gmail.com>