fix: use private identity in realtime (CT-000) (#7436)
Based on this investigation: https://www.notion.so/voiceflow/Realtime-Auth-Spam-2f97c3d20eb84f498b8a0c7922cec8fc
we need to reduce the number of calls from `realtime` to `auth` to not overload `auth` on reload. 

* Instead of calling creator-api to fetch all projects for a workspace, use the ProjectORM in realtime instead.
* Instead of calling public identity, call the private identity endpoints. I've injected the `identityClient` as the `identity` service for this.

By my rough calculation, this reduces 9 calls to auth when loading a canvas.

`hasSource` and `betaFlag` are never used anywhere. I've also had to be a bit of casting, but the return signatures are exactly the same as the non-private endpoints - we've just been casting it on axios before. 
There are a bunch of annoying enum errorsm but this PR would be too big if I try to fix it.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>