fix: cms resources not being cloned (CV3-758) (#7480)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CV3-758**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

Changing assistant publish to clone all CMS resources from the `devVersion` into the new `liveVersion`. 

This is necessary because when publishing a project with Functions with production, the Functions compilation needs to pull the Function definitions from the CMS. However, the existing assistant publish does not clone the Function CMS resources, and so, publish to production will fail because Function compilation fails. 



Co-authored-by: Yauheni <overhawlin@gmail.com>