feat: refactor ml-gateway for LLMs (ACT-024) (#7464)
completely refactored `ml-gateway` to use `nest-js`.
No longer use `ml-sdk`, but use `sdk-http-ml-gateway` instead.

The only thing left from the NLUM is the intent clarity score - we are still preserving that for now.

All prompts have been updated for GPT 3.5 instead of GPT 3.

The only danger is if the user is on a older version of the client, they will be trying to hit the old endpoints, which will fail. Not a big deal but worth pointing out.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>