fix: log project list error (BUG-690) (#7615)
Some customers have seen their project-lists get reset entirely back to "default list".

This is on the legacy dashboard. It's unfortunate but we have some enterprise customers still depedendent on it.

I honestly don't know how to fix the bug because it is impossible to reproduce. I'm trying to add a log where I suspect it is happening. 

Co-authored-by: Tyler Han <tylerhan97@gmail.com>