fix: future-proof the default code (CV3-000) (#7656)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CV3-000**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

We would like to extend the argument object passed into function code with additional properties in the future, such as a `context` object that automatically passes built-in variables.

The issue is that the functions editor does not allow users to edit the first line and we are using object destructuring on the argument objects in the first line.

If we added a new property to the argument object, then users would not be able to gain access to this new property, because they would need to add another property to the object destructuring statement on the first line. This means all code written before the argument object was updated has no access to the new property.

To future-proof functions code, we rework the default functions code by performing the object destructuring on the second line.