fix: collaboration qol updates (BUG-747) (#7748)
This isn't a fix for eltropy's issue, but an attempt to add more logs so we can debug what is exactly causing the reload.

During my investigation I noted that sometimes there's an edge case where the client connects before we listen for the connected message, so the user gets stuck in an infinite "loading collaboration" screen. This is because the actual client is initialized on app load, and the gate might not be rendered by then:
https://github.com/voiceflow/creator-app/blob/164437761886352a82981aec60ca674db2155076/apps/creator-app/src/client/realtime.ts#L16-L24

<img width="719" alt="Screenshot 2024-02-03 at 12 06 46 PM" src="https://github.com/voiceflow/creator-app/assets/5643574/f8b70bb8-67d9-40d4-a5af-13ce2faa60c8">

The fix is very easy.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>