fix: allow api-key export json (BUG-773) (#7809)
The creator-api export VF file is now out of date and many clients rely on it.

Currently there is no way to use the API key with it. The infra team is splitting out the HTTP service from the WSS one, so there should be less load. We should tell users to move over.
https://github.com/voiceflow/platform/blob/708bc57f8bb71ba459f1bbb9ff4f1d106c9a3c1b/libs/sdk-auth/src/platform/nestjs/decorators/user-id.decorator.ts#L9-L20

Co-authored-by: Tyler Han <tylerhan97@gmail.com>