fix: unsupported kb model handling (COR-1050) (#7805)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements COR-1050**

### Brief description. What is this change?

when opening cms kb for a project, if it has an unsupported or invalid model in the summarization settings then the settings modal will crash. This fix checks if a model is not in the supported list, and sets it to default model if needed, also updating this on the backend.