fix: do not load nlu cms for migration [bugfix] (BUG-1166) (#7826)
Noting that KBC has over 100k utterances as part of their NLU model.

When we try to run a migration and load in all CMS items, `realtime` runs out of memory and crashes. This is extremely bad because it crashes not just for them, but for anyone else on that node.
I've confirmed this was the behavior by looking at infrastructure container logs during their session.

The solution for now is to not load in any NLU items during migration. This is still okay because from v4.07 to v5, we are only adding new nlu entries, and not modifying any existing ones.

This way if the user is migrating from v5 to v6/6.01 it won't crash when we try to load all cms items.

full report:
https://www.notion.so/voiceflow/CMS-Overload-f002758425084ac8a95bfca64ceb219e

Co-authored-by: Tyler Han <tylerhan97@gmail.com>