feat: enable html tags on markdown (COR-1722) (#7903)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements COR-1722**

### Brief description. What is this change?

Markdown doesn't natively support underlining. The solution we have is to use the `<u>` tag which requires the rehypeRaw plugin to allow html tags in markdown. This plugin is used in the react-chat library but we aren't using it here which is causing the underlining to not be displayed. Adding the plugin solves this issue.