fix: unnecessary score for folder names (COR-1716) (#7901)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements COR-1716**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

Change the type interface for `CMSTableNameCell` to decouple the name transforms for an item vs. a folder. The original couple accepts a single transform function for both items and folders, leading to an awkward situation where either variables and folders will both automatically replace spaces with underscores or neither or them will. 

The correct data modelling is to provide separate transforms for folders and items.