fix: functions dto accept string (CT-000) (#7931)
We don't actually use `FunctionCompiledVariableDeclarationDTO` or any of it's downstream stuff anywhere in creator-app I believe. We use it purely for types.

We should be careful when modifying DTOs

But we should patch this because the moment someone uses it all functions will break unless we run a migration. All functions today still use `type: "string"`

Related PR:
https://github.com/voiceflow/general-runtime/pull/743

Co-authored-by: Tyler Han <tylerhan97@gmail.com>