fix: disbale create button on creation (COR-1863) (#7961)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements COR-1863**

### Brief description. What is this change?

Currently on project creation the create button is not disabled while creation, which allows the user to spam the button and create multiple projects and can cause the user to dropped as a viewer into a project. Using modal prevent close and disabled state with the button prevents both the issues from occurring.