fix: test modal value regression (COR-000) [bugfix] (#7996)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements COR-000**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

The backend is supposed to send back a `string` or a `{ path: string }` object from `defaultTo`, but this was not specified properly in the backend. This may have misled the frontend developer to only handle the `{ path: string }`. As a result, sending a runtime command where `defaultTo` is a `string`, doesn't get rendered properly by the Test Function modal.

This PR fixes that issue.