fix: laggy javascript editor (COR-1927) (#8001)
Reference comments.

We were not properly deleting the "completers" and it was causing massive lag as it accumulated. Look at behavior below: (this is on current prod) Notice how there are a lot of duplicates, we were running multiple completers at the same time.
<video src="https://github.com/voiceflow/creator-app/assets/5643574/d36791a1-9870-4c1f-b712-f60dcc3dd35a"/>

it was bad hook management from the very original implementation. This exponentially compounds with more variables. 
n variables, m completers, so It's O(mn) 

This was also a very minor memory leak.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>