feat: revert entities and query to required and refactor request object expectations (PL-958) (#8046)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-958**

### Brief description. What is this change?

<!-- Build up some context for your teammates on the changes made here and potential tradeoffs made and/or highlight any topics for discussion -->

The `@voiceflow/dtos` package changed the `query` and `entities` property of the `IntentRequest` type to be optional. This is because the DM API does not require that the user provide `query` or `entities` when they directly send a request object. The use-case for when this occurs is when a user uses a custom LLM to match intents, rather than using VFNLU.

However, changing the type to be optional is problematic because the `general-runtime` was built under the assumption that `query` and `entities` was required. This PR reverts that change by providing a default value. In the `general-runtime`, we must execute the DTO to parse the incoming request, which will automatically set `query` and `entities` to the default value.