fix: remove conditional cloning diagrams (COR-2025) (#8079)
# production race condition
This came because of a race condition/edge case for a user, where their version got deleted, but never recreated. And there was not an invalid reference to `liveVersion` on the project.

This is also very bad for production, because `general-runtime` also uses version, so that means there is a period of time where a project "breaks" because version is deleted. I switched this to upsert instead.

### manual testing
- publish multiple versions ✅
- upload VF file ✅
- duplicate project ✅
- duplicate project link ✅

# removing clone endpoint
The `POST '/v1alpha1/private/environment/:environmentID/clone'` endpoint isn't used at all.

I double checked all of `creator-app` and `general-service`. 
I believe we originally made it for the intents and entities CMS, for `general-service`, but we created 
`/v1alpha1/private/assistant/:assistantID/publish` instead. This is the original JIRA ticket:
<img width="1063" alt="Screenshot 2024-05-07 at 10 10 46 PM" src="https://github.com/voiceflow/creator-app/assets/5643574/f5716282-22a8-4a8a-b4bc-ecb45c4a21c6">

On datadog it is showed as not used in the last few days:
<img width="1920" alt="Screenshot 2024-05-07 at 9 49 49 PM" src="https://github.com/voiceflow/creator-app/assets/5643574/07cfdcf7-140a-4564-806d-77b4dd07ae8d">
You can see the other private routes being used:
<img width="1180" alt="Screenshot 2024-05-07 at 10 11 26 PM" src="https://github.com/voiceflow/creator-app/assets/5643574/be3e9c47-55c4-4b97-8aa1-2a461b1930d6">

Every other usage of `EnvironmentService.clone` has `cloneDiagrams: true` hardcoded.

Co-authored-by: Tyler <tyler@Tylers-MacBook-Pro.local>