fix: ensure husky is installed (ENG-108) (#8257)
we should move the default husky script from:
`"prepare": "husky"`, to
`"postinstall": "husky install",`
the problem I've seen is unless a dev explicitly runs yarn husky or npx husky install husky will just not work in a repo. You can recreate this by deleting the platform repo and clone it again, and running a commit. The problem is for any new people, husky will not be installed properly and no linting on commit is run.

Yeah prepare is not a supported lifecycle on yarn but it is for npm:
https://yarnpkg.com/advanced/lifecycle-scripts
https://docs.npmjs.com/cli/v10/using-npm/scripts#life-cycle-scripts

It's a very small overhead on install:
<img width="693" alt="Screenshot 2024-06-10 at 12 03 31 PM" src="https://github.com/voiceflow/platform/assets/5643574/76901a5a-3855-48e0-a8d3-7ae6a5c5c1df">



Co-authored-by: Tyler <tylerhan97@gmail.com>