chore: vulnerability update for protobufjs (COR-2200) (#8199)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements COR-2200**

### Brief description. What is this change?

According to [this](https://github.com/advisories/GHSA-h755-8qp9-cq85) there is a prototype pollution vulnerabiliity in the library protobufjs v7.2.3  .     
This is a dependency of google-gax which is used by `google-cloud/pubsub` and `google-cloud/firestore`.  
In both these cases, google made a minor version upgrade to protobuf v7.2.4 and [claimed](https://github.com/googleapis/gax-nodejs/releases/tag/v3.6.1) that this solves the vulnerabilty. It does not though! (I opened a ticket to google reporting that, but I wouldn't hold my breath).  

That is why I bumped them by major versions.  
In the changelog the only breaking change for pubsub was bumping node version, and we already use a newer node version. For firestore there was another breaking change in a specific method call, which didn't seem relevant to our use case.

BUT, i feel like i don't know how to properly test this, since it's in `ml-gateway` and not familiar yet with that (Hence, the `dont merge` label).


### Checklist

- [ ] changes have been validated in an ephemeral environment
- [ ] this is a breaking change and should publish a new major version
- [ ] appropriate tests have been written
- [ ] any new env vars have been added to the [notion doc](https://www.notion.so/voiceflow/Add-Environment-Variables-be1b0136479f45f1adece7995a7adbfb) and infra has been notified